Handling CreateProcess myself. Cropped screenshot.
Subprocess only partially implements STARTUPINFO(). It seems to ignore
lpTitle and dwFillAttribute. Implementing my own wrapper for it.

On Windows XP and Server 2012, cmd.exe does not fully maximize the
entire window. It only vertically maximizes. This may lead to false
positives on matching subimages on the screen. Mitigating this by doing
cropped screenshots with the help of GetWindowRect().